backend/requirements.txt
torch
pandas
scikit-learn
requests
numpy
python-dotenv==1.0.1
fastapi
uvicorn
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2 